Create compose util, general util cleanup
